Introducing 'b2tinnt_fchin_in_core' switch to improve convergence of cases with drifts.
The recent problems with ITER runs convergence with drifts are as follows. The neoclassical solution for electric field inside the separatrix demands very small coefficient in front of artificial anomalous conductivity: csig = 1e-7. In this case even small change of the current through the flux surface at the time step produces big jump in the electric field at the same step. The ion-neutral friction current fchin is a fluctuating value in case of EIRENE neutrals. Its fluctuations in our case are of the order of 1A after integration over the flux surface. It produces fluctuations of electric field of the order of of dE=1-2 kV/m. The corresponding change of ExB drift and therefore of density perturbation on the flux surface make the solution oscillate significantly around converged value. Therefore it is necessary for convergence of most of ITER runs to turn off this current inside the separatrix. At the same time, the absolute time-averaged value of this current (about 3 A) is much smaller than that of viscosity and diamagnetic currents, and it does not influence solution significantly. The estimate can be actually made dE/E~(dfchin/fchin)*(fchin/fchan). If variation of fchin is of the order of fchin itself, then even the small current fchin of the order of anomalous current fchan makes electric field jump considerably.